+++ mitra_cli/src/commands/mod.rs
@@ -4,4 +4,5 @@ pub mod config;
 pub mod filter;
 pub mod invoice;
 pub mod post;
+pub mod process;
 pub mod storage;